finished router and add roadpage.js
