+++ src/component/Footer.css
@@ -2,5 +2,5 @@
     background-color: rgb(56, 147, 147);
     height: 6vh;
     width: 100%;
-    position: relative;
+    position: absolute;
 }
