+++ src/component/Header.css
@@ -1,6 +1,6 @@
 .Header {
-    height: 92vh;
-    width: 100%;
+    height: 88vh;
+    width: 100vw;
     padding: 0px;
     display: flex;
     align-items: center;
@@ -39,7 +39,8 @@
     width: 100%;
     padding: 0px;
     margin-bottom: -20px;
-    border:3px solid #FA7F72;
+    border-top:3px solid #FA7F72;
+    border-bottom: 3px solid #FA7F72 ;
 }
 
 .RequestForm{
