+++ src/component/Navbar.css
@@ -8,7 +8,8 @@
 }
 
 * {
-    margin: 0%;
+    margin: 0;
+    padding: 0;
     
 }
