+++ src/index.js
@@ -3,10 +3,12 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
+import {BrowserRouter as Router} from 'react-router-dom';
 
 ReactDOM.render(
-  
+  <Router>
     <App />
+  </Router>
   ,
   document.getElementById('root')
 );
